+++ app/components/Views/confirmations/components/info/contract-interaction/contract-interaction.tsx
@@ -6,10 +6,10 @@ import { SimulationDetails } from '../../../../../UI/SimulationDetails/Simulatio
 import { useConfirmationMetricEvents } from '../../../hooks/metrics/useConfirmationMetricEvents';
 import { use7702TransactionType } from '../../../hooks/7702/use7702TransactionType';
 import { useTransactionMetadataRequest } from '../../../hooks/transactions/useTransactionMetadataRequest';
-import AccountNetworkInfo from '../../rows/account-network-info-row';
+import AccountNetworkInfoRow from '../../rows/account-network-info-row';
 import OriginRow from '../../rows/origin-row';
 import AdvancedDetailsRow from '../../rows/transactions/advanced-details-row/advanced-details-row';
-import GasFeesDetails from '../../rows/transactions/gas-fee-details';
+import GasFeesDetailsRow from '../../rows/transactions/gas-fee-details-row';
 import SwitchAccountTypeInfoRow from '../../rows/switch-account-type-info-row';
 
 const ContractInteraction = () => {
@@ -22,7 +22,7 @@ const ContractInteraction = () => {
 
   return (
     <View>
-      <AccountNetworkInfo />
+      <AccountNetworkInfoRow />
       {isBatchedUpgrade && <SwitchAccountTypeInfoRow />}
       <SimulationDetails
         transaction={transactionMetadata as TransactionMeta}
@@ -30,7 +30,7 @@ const ContractInteraction = () => {
         isTransactionsRedesign
       />
       <OriginRow isSignatureRequest={false} />
-      <GasFeesDetails />
+      <GasFeesDetailsRow />
       <AdvancedDetailsRow />
     </View>
   );
