+++ app/components/Views/confirmations/components/info/typed-sign-v3v4/typed-sign-v3v4.tsx
@@ -1,12 +1,12 @@
 import React from 'react';
-import AccountNetworkInfo from '../../rows/account-network-info-row';
+import AccountNetworkInfoRow from '../../rows/account-network-info-row';
 import { InfoSectionOriginAndDetails } from './info-section-origin-and-details';
 import Message from './message';
 import TypedSignV3V4Simulation from './simulation';
 
 const TypedSignV3V4 = () => (
     <>
-      <AccountNetworkInfo />
+      <AccountNetworkInfoRow />
       <TypedSignV3V4Simulation />
       <InfoSectionOriginAndDetails />
       <Message />
