+++ app/components/Views/confirmations/components/rows/transactions/advanced-details-row/index.ts
@@ -1,0 +1,1 @@
+export { default } from './advanced-details-row';
