+++ app/components/Views/confirmations/components/rows/transactions/network-and-origin-row/index.ts
@@ -1,0 +1,1 @@
+export * from './network-and-origin-row';
