--- app/components/Views/confirmations/components/rows/transactions/network-row/index.ts
@@ -1,1 +1,0 @@
-export { default } from './network-row';
