+++ app/components/Views/confirmations/external/staking/info/staking-claim/staking-claim.tsx
@@ -13,7 +13,7 @@ import { useTransactionMetadataRequest } from '../../../../hooks/transactions/us
 import InfoSection from '../../../../components/UI/info-row/info-section';
 import StakingContractInteractionDetails from '../../components/staking-contract-interaction-details/staking-contract-interaction-details';
 import { HeroRow } from '../../../../components/rows/transactions/hero-row';
-import GasFeesDetails from '../../../../components/rows/transactions/gas-fee-details';
+import GasFeesDetailsRow from '../../../../components/rows/transactions/gas-fee-details-row';
 
 const StakingClaim = ({
   route,
@@ -55,7 +55,7 @@ const StakingClaim = ({
       <InfoSection>
         <StakingContractInteractionDetails />
       </InfoSection>
-      <GasFeesDetails disableUpdate />
+      <GasFeesDetailsRow disableUpdate />
     </>
   );
 };
